% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bird}
\alias{bird}
\title{bird data:}
\format{A list with 3 elements:
\describe{
  \item{tre}{phy tree information, character}
  \item{abun}{abundance data, data frame}
  \item{inci}{incidenece data, list}
  ...
}}
\usage{
bird
}
\description{
A list containing phy tree information, abun, inci
diamonds.
}
\keyword{datasets}
