% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_p.r
\name{Boot_p}
\alias{Boot_p}
\title{R code for JADE (Joint species-rank Abundance Distribution/Estimation): two parameters &
2013 Entropy and the species accumulation curve Appendix S2: Estimating species relative abundance :one parameter
based on individual-based (abundance) and sampling-unit-based (incidence) data.}
\usage{
Boot_p(x, zero = TRUE, Bootype = "One", datatype = "abundance")
}
\arguments{
\item{x}{a vector/matrix/list of species abundance frequency}

\item{zero}{reserves zero frequency or not. Default is TRUE.}

\item{Bootype}{a string= "JADE" or "One". Default is "One"}

\item{datatype}{a string}
}
\value{
a vector/matrix/list with species relative abundance or detection probability distribution
}
\description{
R code for JADE (Joint species-rank Abundance Distribution/Estimation): two parameters &
2013 Entropy and the species accumulation curve Appendix S2: Estimating species relative abundance :one parameter
based on individual-based (abundance) and sampling-unit-based (incidence) data.
}
\examples{
data(bird)
bird.inc <- bird$inci
bird.abun<- bird$abun
Boot_p(bird$abun, datatype="abundance")
Boot_p(bird$abun, datatype="incidence_raw")
}
