% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_p.r
\name{bootp_jade_abu}
\alias{bootp_jade_abu}
\title{DetAbu(x, zero=FALSE) is a function of estimating detected species relative abundance.}
\usage{
bootp_jade_abu(Spec, zero = TRUE)
}
\arguments{
\item{Spec}{a vector of species abundance frequency}

\item{zero}{reserves zero frequency or not. Default is FALSE.}
}
\value{
a numerical vector and its length is the same as Spec
bootp_jade_abu(Spec,zero) is a function for abundance data
bootp_jade_inc(Spec,zero) is a function for incidence data
}
\description{
DetAbu(x, zero=FALSE) is a function of estimating detected species relative abundance.
}
