% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_p.r
\name{bootp_one_abu}
\alias{bootp_one_abu}
\title{Estimating species relative abundance/incidnece bootstrap Pi
Chao, A., Wang, Y. T., and Jost, L. (2013).
Entropy and the species accumulation curve: a novel estimator of entropy via discovery rates of new species.
Appendix S2:  Estimating species relative abundance :one parameter
bootp_one_abu(x) is a function for abundance data
bootp_one_inc(x) is a function for incidence data}
\usage{
bootp_one_abu(Spec, zero = TRUE)
}
\arguments{
\item{Spec}{a vector of species abundance frequency}

\item{zero}{reserves zero frequency or not. Default is TRUE.}
}
\value{
a numerical vector and its length is the same as Spec
}
\description{
Estimating species relative abundance/incidnece bootstrap Pi
Chao, A., Wang, Y. T., and Jost, L. (2013).
Entropy and the species accumulation curve: a novel estimator of entropy via discovery rates of new species.
Appendix S2:  Estimating species relative abundance :one parameter
bootp_one_abu(x) is a function for abundance data
bootp_one_inc(x) is a function for incidence data
}
