% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_p.r
\name{bootp_one_abu_sar}
\alias{bootp_one_abu_sar}
\title{estimate the bootstrap population of an incidence sample which was drawn without replacement.
Chao, A.,... (2019).
bootp_one_abu_sar(x) is a function for abundance data
bootp_one_inc_sar(x) is a function for incidence data}
\usage{
bootp_one_abu_sar(Spec, zero = TRUE, rho = NaN)
}
\arguments{
\item{Spec}{a vector of species abundance frequency}

\item{zero}{reserves zero frequency or not. Default is TRUE.}

\item{rho}{a numerical value as sampling fraction}
}
\value{
a numerical vector and its length is the same as Spec
}
\description{
estimate the bootstrap population of an incidence sample which was drawn without replacement.
Chao, A.,... (2019).
bootp_one_abu_sar(x) is a function for abundance data
bootp_one_inc_sar(x) is a function for incidence data
}
\examples{
data(bird)
bird.abun<- bird$abun
Boot_p(bird.abun,zero=FALSE,Bootype="SAR",datatype="abundance",rho=0.7)
data(incdata)
Boot_p(incdata,zero=FALSE,Bootype="SAR",datatype="incidence",rho=0.7)
}
