% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkType.r
\name{check_Bootype}
\alias{check_Bootype}
\title{R code for check bootstrap type}
\usage{
check_Bootype(Bootype)
}
\arguments{
\item{Bootype}{a string "One", "JADE"}
}
\value{
a correct Bootype or stop
}
\description{
R code for check bootstrap type
}
