% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkType.r
\name{check_data}
\alias{check_data}
\title{R code for check data}
\usage{
check_data(x, datatype = "abundance")
}
\arguments{
\item{x}{a vector}

\item{datatype}{a string}
}
\value{
if correct no return, otherwise stop
}
\description{
R code for check data
}
