% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkType.r
\name{check_datatype}
\alias{check_datatype}
\title{R code for check datatype}
\usage{
check_datatype(datatype)
}
\arguments{
\item{datatype}{a string "abundance", "incidence", "incidence_freq", "incidence_raw"}
}
\value{
a correct datatype or stop
}
\description{
R code for check datatype
}
