% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_p.r
\name{DetAbu}
\alias{DetAbu}
\title{DetAbu(x, zero=FALSE) is a function of estimating detected species relative abundance.}
\usage{
DetAbu(x, zero = FALSE)
}
\arguments{
\item{x}{a vector of species abundance frequency}

\item{zero}{reserves zero frequency or not. Default is FALSE.}
}
\value{
a numerical vector
}
\description{
DetAbu(x, zero=FALSE) is a function of estimating detected species relative abundance.
}
