% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyBranchAL.r
\name{phy_BranchAL_Abu}
\alias{phy_BranchAL_Abu}
\title{R code for phylo to Chaophyabu function, we defind chaophytree objects}
\usage{
phy_BranchAL_Abu(phylo, data, datatype = "abundance", refT = 0,
  rootExtend = T, remove0 = T)
}
\arguments{
\item{phylo}{a phylo object}

\item{data}{a vector with names}
}
\value{
a Chaophyabu objects
}
\description{
R code for phylo to Chaophyabu function, we defind chaophytree objects
}
\examples{
data(AbuALdata)
adata<-AbuALdata$abudata
atree<-AbuALdata$tree
vdata<-adata$EM
names(vdata)<-rownames(adata)
refTs<-c(400,325,250)
result<-phy_BranchAL_Abu(atree,vdata,datatype="abundance",refTs)
result$treeNabu
result$treeH
result$BLbyT


}
