% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyBranchALinc.r
\name{phy_BranchAL_IncBootP}
\alias{phy_BranchAL_IncBootP}
\title{R code for phylo to Chaophyabu function, calculte inodeP by incidence bootstrap p
the formula is 1-(1-p1)(1-p2)}
\usage{
phy_BranchAL_IncBootP(phylo, pdata, refT = 0, rootExtend = T,
  remove0 = T)
}
\arguments{
\item{phylo}{a phylo object}

\item{pdata}{a vector with names: incidence bootstrap p}
}
\value{
a Chaophyabu objects
}
\description{
R code for phylo to Chaophyabu function, calculte inodeP by incidence bootstrap p
the formula is 1-(1-p1)(1-p2)
}
\examples{
data(phyincPdata)
pdata <- phyincPdata$pdata
phylotree <- phyincPdata$tree
refTs<-c(80,90,100)
result<-phy_BranchAL_IncBootP(phylotree,pdata,refTs,remove0=FALSE)
result$treeNincBP
result$treeH
result$BLbyT


}
