% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{phybird}
\alias{phybird}
\title{phybird data:}
\format{A list with 5 elements:
\describe{
  \item{tre}{phy tree information, character}
  \item{abun}{abundance data, data frame}
  \item{inci}{incidenece data, list}
  \item{phylo}{phylo tree information, phylo}
  \item{chaophytree}{tree and abundance combined data, tibble}
  ...
}}
\usage{
phybird
}
\description{
A list containing phy tree information, abun, inci
diamonds.
}
\keyword{datasets}
