% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyBranchAL.r
\name{phyBranchAL_Abu}
\alias{phyBranchAL_Abu}
\title{R code for phylo to Chaophyabu function, speed up performance}
\usage{
phyBranchAL_Abu(phylo, data, datatype = "abundance", refT = 0,
  rootExtend = T, removeAbu0 = T)
}
\arguments{
\item{phylo}{a phylo object}

\item{data}{a vector with names}
}
\value{
a Chaophyabu objects
}
\description{
R code for phylo to Chaophyabu function, speed up performance
}
\examples{
data(AbuALdata)
adata<-AbuALdata$abudata
atree<-AbuALdata$tree
vdata<-adata$EM
names(vdata)<-rownames(adata)
refTs<-c(400,325,250)
result<-phyBranchAL_Abu(atree,vdata,datatype="abundance",refTs)
result$treeNabu
result$treeH
result$BLbyT


}
