% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyBranchALinc.r
\name{phyBranchAL_Inc}
\alias{phyBranchAL_Inc}
\title{R code for phylo to Chaophyabu function, speed up performance}
\usage{
phyBranchAL_Inc(phylo, data, datatype = "incidence_raw", refT = 0,
  rootExtend = T, remove0 = T)
}
\arguments{
\item{phylo}{a phylo object}

\item{data}{a vector with names}
}
\value{
a Chaophyabu objects
}
\description{
R code for phylo to Chaophyabu function, speed up performance
}
\examples{
data(phybird.new)
bird.inc <- phybird.new$inci
bird.lab <- rownames(phybird.new$abun)
bird.phy <- phybird.new$chaophytree
tree<-tidytree::as.phylo(bird.phy$phytree)
Idata<-bird.inc$North.site
refTs<-c(80,90,100)
result<-phyBranchAL_Abu(tree,Idata,datatype="incidence_raw",refTs)
result$treeNabu
result$treeH
result$BLbyT


}
