% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyExpandData.r
\name{phyExpandData}
\alias{phyExpandData}
\title{Expand branch abundance/incience and branch length}
\usage{
phyExpandData(x, labels, phy, datatype = "abundance")
}
\arguments{
\item{x}{a vector/matrix/list of species abundances or a matrix of raw incidence table.\cr}

\item{labels}{a vector of species name for input data.\cr}

\item{phy}{a phylogenetic tree with \code{"chaophytree"} class.\cr}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}), species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}).}
}
\value{
a tibble tree with abundances.
}
\description{
\code{phyExpandData}: Expand branch abundance/incience and branch length
}
\examples{
data(phybird.new)
bird.abu <- phybird.new$abun
bird.inc <- phybird.new$inci
bird.lab <- rownames(phybird.new$abun)
bird.phy <- phybird.new$chaophytree
phyExpandData(bird.abu, labels=bird.lab, phy=bird.phy, datatype="abundance")
phyExpandData(bird.inc, labels=bird.lab, phy=bird.phy, datatype="incidence_raw")

}
