% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{phyincisimple.new}
\alias{phyincisimple.new}
\title{phyincisimple data:}
\format{A list with 5 elements:
\describe{
  \item{inci.simple.data}{incidence data, matrix}
  \item{tree.simple}{phylo tree information, phylo}
  \item{tree.simple.phytree}{chaophytree data, chaophytree: new format}

  ...
}}
\usage{
phyincisimple.new

phyincisimple.new
}
\description{
A list containing inci,phy tree information
diamonds.

A list containing inci,phy tree information
diamonds.
}
\keyword{datasets}
