% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{phyincPdata}
\alias{phyincPdata}
\title{phyBranchALinc.r data:}
\format{A list with 2 elements:
\describe{
  \item{pdata}{abundance data, vector}
  \item{phylotree}{phylo tree information, phylo}


  ...
}}
\usage{
phyincPdata
}
\description{
A list containing pdata,phylo tree information
diamonds.
}
\keyword{datasets}
