% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyLengthbyT.r
\name{phyLengthbyT}
\alias{phyLengthbyT}
\title{Calculate branch length by reference ageT}
\usage{
phyLengthbyT(Ts, phy, datatype = "abundance")
}
\arguments{
\item{phy}{a phylogenetic tree with \code{"chaophytree"} class.\cr}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}), species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}).}

\item{Tx}{a numerical vector of ageT.\cr}
}
\value{
a maxtrix with new branch length by ageT(Tx).
}
\description{
\code{phyLengthbyT}: Calculate branch length by reference ageT
}
\examples{
data(phybird)
bird.abu <- phybird$abun
bird.inc <- phybird$inci
bird.lab <- rownames(phybird$abun)
bird.phy <- phybird$chaophytree
phyLengthbyT(Ts=c(75,55),  phy=bird.phy, datatype="abundance")

}
