% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo2chaotools.r
\name{phylo2chaolabphy}
\alias{phylo2chaolabphy}
\title{R code for phylo to phylog like function, we defind chaolabphy objects}
\usage{
phylo2chaolabphy(phylo)
}
\arguments{
\item{phylo}{a phylo object}
}
\value{
a chaolabphy objects
}
\description{
R code for phylo to phylog like function, we defind chaolabphy objects
}
\examples{
data(treesample)
newphy<-phylo2chaolabphy(treesample)
leaves <- newphy$leaves
nodes<- newphy$nodes
parts<- newphy$parts

}
