% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo2chaotools.r
\name{phylo2phytree}
\alias{phylo2phytree}
\title{R code for phylo to phytree function, we defind chaophytree objects}
\usage{
phylo2phytree(phylo)
}
\arguments{
\item{phylo}{a phylo object}
}
\value{
a chaophytree objects
}
\description{
R code for phylo to phytree function, we defind chaophytree objects
}
\examples{
data(treesample)
newphy<-phylo2phytree(treesample)
leaves <- newphy$leaves
nodes<- newphy$nodes
treedata<-newphy$phytree
treeH<-newphy$treeH


}
