% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_p.r
\name{UndAbu}
\alias{UndAbu}
\title{Estimating undetected species relative abundance}
\usage{
UndAbu(x)
}
\arguments{
\item{x}{a vector of species abundance frequency}
}
\value{
a numerical vector
}
\description{
\code{UndAbu} Estimating undetected species relative abundance
}
